Moving logic to add click into player.ts.
